+++ src/pipecat/transports/base_output.py
@@ -208,12 +208,14 @@ class BaseOutputTransport(FrameProcessor):
     async def _bot_started_speaking(self):
         if not self._bot_speaking:
             logger.debug("Bot started speaking")
+            await self.push_frame(BotStartedSpeakingFrame())
             await self.push_frame(BotStartedSpeakingFrame(), FrameDirection.UPSTREAM)
             self._bot_speaking = True
 
     async def _bot_stopped_speaking(self):
         if self._bot_speaking:
             logger.debug("Bot stopped speaking")
+            await self.push_frame(BotStoppedSpeakingFrame())
             await self.push_frame(BotStoppedSpeakingFrame(), FrameDirection.UPSTREAM)
             self._bot_speaking = False
 
@@ -323,7 +325,7 @@ class BaseOutputTransport(FrameProcessor):
     #
 
     async def send_image(self, frame: OutputImageRawFrame | SpriteFrame):
-        await self.process_frame(frame, FrameDirection.DOWNSTREAM)
+        await self.queue_frame(frame, FrameDirection.DOWNSTREAM)
 
     async def _draw_image(self, frame: OutputImageRawFrame):
         desired_size = (self._params.camera_out_width, self._params.camera_out_height)
@@ -394,7 +396,7 @@ class BaseOutputTransport(FrameProcessor):
     #
 
     async def send_audio(self, frame: OutputAudioRawFrame):
-        await self.process_frame(frame, FrameDirection.DOWNSTREAM)
+        await self.queue_frame(frame, FrameDirection.DOWNSTREAM)
 
     def _next_audio_frame(self) -> AsyncGenerator[AudioRawFrame, None]:
         async def without_mixer(vad_stop_secs: float) -> AsyncGenerator[AudioRawFrame, None]:
@@ -452,6 +454,7 @@ class BaseOutputTransport(FrameProcessor):
                 # it's actually speaking.
                 if isinstance(frame, TTSAudioRawFrame):
                     await self._bot_started_speaking()
+                    await self.push_frame(BotSpeakingFrame())
                     await self.push_frame(BotSpeakingFrame(), FrameDirection.UPSTREAM)
 
                 # Also, push frame downstream in case anyone else needs it.